limit max client body size

+++ config/ServerConfig.hpp
@@ -9,7 +9,7 @@ public:
     std::string host;
     std::string server_name;
     std::map<int, std::string> error_page;
-    int client_max_body_size;
+    long unsigned int client_max_body_size;
     std::string index;
     std::string root;
     std::map<std::string, LocationConfig> locations;
